Moved Listing from search model to shared model
