+++ app/src/main/java/com/sthoray/allright/data/model/search/SearchResponse.kt
@@ -1,5 +1,5 @@
 package com.sthoray.allright.data.model.search
-
+import com.sthoray.allright.data.model.shared.Listing
 /**
  * Model for search responses.
  *
