+++ app/src/main/java/com/sthoray/allright/ui/search/adapter/SearchAdapter.kt
@@ -8,7 +8,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import coil.api.load
 import com.sthoray.allright.R
-import com.sthoray.allright.data.model.search.Listing
+import com.sthoray.allright.data.model.shared.Listing
 import kotlinx.android.synthetic.main.item_layout_search.view.*
 
 /**
